fix(posts): Temporarily fixes image referencing issue. See body for explanation.
For some reason, Hugo can't find the static folder, nor use page bundles. In both states, it's locally able to find the image files, the index.html that's generated finds the local files, but as soon as I use hugo server, it fails to find it. I'm not sure why this is happening, nor the solution, but for now this is the best I can do. I must discover a solution, as this is very much a workaround solution.